+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## [1.2.1]
+
+### Fixed
+
+- absolute or relative path fixed
+- caching fixed in general
+
 ## [1.2.0]
 
 ### Added